@@ -11,23 +11,20 @@
             "homepage": "https://github.com/jnvsor"
         },
         {
-            "name": "Rokas Šleinius",
-            "homepage": "https://github.com/raveren"
-        },
-        {
             "name": "Contributors",
             "homepage": "https://github.com/kint-php/kint/graphs/contributors"
         }
     ],
     "require": {
-        "php": ">=5.3.6"
+        "php": ">=5.6"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.0",
-        "phpunit/phpunit": "^4.0",
-        "symfony/finder": "^2.0 || ^3.0 || ^4.0",
+        "friendsofphp/php-cs-fixer": "^3.0",
+        "phpunit/phpunit": "^9.0",
+        "phpspec/prophecy-phpunit": "^2",
+        "symfony/finder": "^3.0 || ^4.0 || ^5.0",
         "seld/phar-utils": "^1.0",
-        "vimeo/psalm": "^2.0"
+        "vimeo/psalm": "^4.0"
     },
     "autoload": {
         "files": ["init.php"],
@@ -42,14 +39,11 @@
     },
     "config": {
         "platform": {
-            "php": "7.0"
+            "php": "7.4"
         }
     },
     "scripts": {
-        "post-update-cmd": [
-            "npm ci",
-            "bundler install --deployment"
-        ],
+        "post-update-cmd": "npm ci",
         "post-install-cmd": "@post-update-cmd",
         "clean": [
             "rm -rf resources/compiled/",
@@ -62,25 +56,18 @@
         ],
         "format:php": "php-cs-fixer fix",
         "format:js": "npm run format:js",
-        "format:sass": "bundler exec sass-convert -RiF scss -T scss resources/sass/",
+        "format:sass": "npm run format:sass",
         "build": [
             "@build:sass",
             "@build:js",
             "@build:php"
         ],
-        "build:sass": "bundler exec sass --sourcemap=none --update -fCt compressed resources/sass:resources/compiled",
+        "build:sass": "npm run build:sass",
         "build:js": "npm run build:js",
         "build:php": "php ./build.php",
         "analyze": "psalm --show-info=false"
     },
     "suggest": {
-        "kint-php/kint-twig": "Provides d() and s() functions in twig templates",
-        "kint-php/kint-js": "Provides a simplified dump to console.log()",
-        "ext-mbstring": "Provides string encoding detection",
-        "ext-iconv": "Provides fallback detection for ambiguous legacy string encodings such as the Windows and ISO 8859 code pages",
-        "ext-ctype": "Simple data type tests",
-        "symfony/polyfill-mbstring": "Replacement for ext-mbstring if missing",
-        "symfony/polyfill-iconv": "Replacement for ext-iconv if missing",
-        "symfony/polyfill-ctype": "Replacement for ext-ctype if missing"
+        "kint-php/kint-twig": "Provides d() and s() functions in twig templates"
     }
 }
