@@ -27,7 +27,7 @@ namespace Kint;
 
 class CallFinder
 {
-    private static $ignore = array(
+    private static $ignore = [
         T_CLOSE_TAG => true,
         T_COMMENT => true,
         T_DOC_COMMENT => true,
@@ -35,7 +35,7 @@ class CallFinder
         T_OPEN_TAG => true,
         T_OPEN_TAG_WITH_ECHO => true,
         T_WHITESPACE => true,
-    );
+    ];
 
     /**
      * Things we need to do specially for operator tokens:
@@ -43,17 +43,19 @@ class CallFinder
      * - Wrap the access path in parentheses if there
      *   are any of these in the final short parameter.
      */
-    private static $operator = array(
+    private static $operator = [
         T_AND_EQUAL => true,
         T_BOOLEAN_AND => true,
         T_BOOLEAN_OR => true,
         T_ARRAY_CAST => true,
         T_BOOL_CAST => true,
+        T_CLASS => true,
         T_CLONE => true,
         T_CONCAT_EQUAL => true,
         T_DEC => true,
         T_DIV_EQUAL => true,
         T_DOUBLE_CAST => true,
+        T_FUNCTION => true,
         T_INC => true,
         T_INCLUDE => true,
         T_INCLUDE_ONCE => true,
@@ -84,6 +86,9 @@ class CallFinder
         T_STRING_CAST => true,
         T_UNSET_CAST => true,
         T_XOR_EQUAL => true,
+        T_POW => true,
+        T_POW_EQUAL => true,
+        T_DOUBLE_ARROW => true,
         '!' => true,
         '%' => true,
         '&' => true,
@@ -100,9 +105,9 @@ class CallFinder
         '^' => true,
         '|' => true,
         '~' => true,
-    );
+    ];
 
-    private static $strip = array(
+    private static $strip = [
         '(' => true,
         ')' => true,
         '[' => true,
@@ -112,48 +117,70 @@ class CallFinder
         T_OBJECT_OPERATOR => true,
         T_DOUBLE_COLON => true,
         T_NS_SEPARATOR => true,
-    );
+    ];
+
+    private static $classcalls = [
+        T_DOUBLE_COLON => true,
+        T_OBJECT_OPERATOR => true,
+    ];
+
+    private static $namespace = [
+        T_STRING => true,
+    ];
 
     public static function getFunctionCalls($source, $line, $function)
     {
-        static $up = array(
+        static $up = [
             '(' => true,
             '[' => true,
             '{' => true,
             T_CURLY_OPEN => true,
             T_DOLLAR_OPEN_CURLY_BRACES => true,
-        );
-        static $down = array(
+        ];
+        static $down = [
             ')' => true,
             ']' => true,
             '}' => true,
-        );
-        static $modifiers = array(
+        ];
+        static $modifiers = [
             '!' => true,
             '@' => true,
             '~' => true,
             '+' => true,
             '-' => true,
-        );
-        static $identifier = array(
+        ];
+        static $identifier = [
             T_DOUBLE_COLON => true,
             T_STRING => true,
             T_NS_SEPARATOR => true,
-        );
-
-        if (KINT_PHP56) {
-            self::$operator[T_POW] = true;
-            self::$operator[T_POW_EQUAL] = true;
-        }
+        ];
 
         if (KINT_PHP70) {
             self::$operator[T_SPACESHIP] = true;
         }
 
+        if (KINT_PHP74) {
+            self::$operator[T_COALESCE_EQUAL] = true;
+        }
+
+        if (KINT_PHP80) {
+            $up[T_ATTRIBUTE] = true;
+            self::$operator[T_MATCH] = true;
+            self::$strip[T_NULLSAFE_OBJECT_OPERATOR] = true;
+            self::$classcalls[T_NULLSAFE_OBJECT_OPERATOR] = true;
+            self::$namespace[T_NAME_FULLY_QUALIFIED] = true;
+            self::$namespace[T_NAME_QUALIFIED] = true;
+            self::$namespace[T_NAME_RELATIVE] = true;
+            $identifier[T_NAME_FULLY_QUALIFIED] = true;
+            $identifier[T_NAME_QUALIFIED] = true;
+            $identifier[T_NAME_RELATIVE] = true;
+        }
+
         $tokens = \token_get_all($source);
         $cursor = 1;
-        $function_calls = array();
-        $prev_tokens = array(null, null, null);
+        $function_calls = [];
+        // Performance optimization preventing backwards loops
+        $prev_tokens = [null, null, null];
 
         if (\is_array($function)) {
             $class = \explode('\\', $function[0]);
@@ -183,10 +210,16 @@ class CallFinder
                 continue;
             }
 
-            $prev_tokens = array($prev_tokens[1], $prev_tokens[2], $token);
+            $prev_tokens = [$prev_tokens[1], $prev_tokens[2], $token];
 
             // Check if it's the right type to be the function we're looking for
-            if (T_STRING !== $token[0] || \strtolower($token[1]) !== $function) {
+            if (!isset(self::$namespace[$token[0]])) {
+                continue;
+            }
+
+            $ns = \explode('\\', \strtolower($token[1]));
+
+            if (\end($ns) !== $function) {
                 continue;
             }
 
@@ -198,7 +231,7 @@ class CallFinder
 
             // Check if it matches the signature
             if (null === $class) {
-                if ($prev_tokens[1] && \in_array($prev_tokens[1][0], array(T_DOUBLE_COLON, T_OBJECT_OPERATOR), true)) {
+                if ($prev_tokens[1] && isset(self::$classcalls[$prev_tokens[1][0]])) {
                     continue;
                 }
             } else {
@@ -206,7 +239,15 @@ class CallFinder
                     continue;
                 }
 
-                if (!$prev_tokens[0] || T_STRING !== $prev_tokens[0][0] || \strtolower($prev_tokens[0][1]) !== $class) {
+                if (!$prev_tokens[0] || !isset(self::$namespace[$prev_tokens[0][0]])) {
+                    continue;
+                }
+
+                /** @var array{int, string, int} $prev_tokens[0] */
+                // All self::$namespace tokens are T_ constants
+                $ns = \explode('\\', \strtolower($prev_tokens[0][1]));
+
+                if (\end($ns) !== $class) {
                     continue;
                 }
             }
@@ -216,9 +257,9 @@ class CallFinder
             $offset = $nextReal + 1; // The start of the function call
             $instring = false; // Whether we're in a string or not
             $realtokens = false; // Whether the current scope contains anything meaningful or not
-            $any_realtokens = false; // Whether the current parameter contains anything meaningful or not
-            $params = array(); // All our collected parameters
-            $shortparam = array(); // The short version of the parameter
+            $paramrealtokens = false; // Whether the current parameter contains anything meaningful
+            $params = []; // All our collected parameters
+            $shortparam = []; // The short version of the parameter
             $param_start = $offset; // The distance to the start of the parameter
 
             // Loop through the following tokens until the function call ends
@@ -232,7 +273,7 @@ class CallFinder
                 }
 
                 if (!isset(self::$ignore[$token[0]]) && !isset($down[$token[0]])) {
-                    $any_realtokens = $realtokens = true;
+                    $paramrealtokens = $realtokens = true;
                 }
 
                 // If it's a token that makes us to up a level, increase the depth
@@ -271,11 +312,12 @@ class CallFinder
                     $shortparam[] = '"';
                 } elseif (1 === $depth) {
                     if (',' === $token[0]) {
-                        $params[] = array(
+                        $params[] = [
                             'full' => \array_slice($tokens, $param_start, $offset - $param_start),
                             'short' => $shortparam,
-                        );
-                        $shortparam = array();
+                        ];
+                        $shortparam = [];
+                        $paramrealtokens = false;
                         $param_start = $offset + 1;
                     } elseif (T_CONSTANT_ENCAPSED_STRING === $token[0] && \strlen($token[1]) > 2) {
                         $shortparam[] = $token[1][0].'...'.$token[1][0];
@@ -286,11 +328,11 @@ class CallFinder
 
                 // Depth has dropped to 0 (So we've hit the closing paren)
                 if ($depth <= 0) {
-                    if ($any_realtokens) {
-                        $params[] = array(
+                    if ($paramrealtokens) {
+                        $params[] = [
                             'full' => \array_slice($tokens, $param_start, $offset - $param_start),
                             'short' => $shortparam,
-                        );
+                        ];
                     }
 
                     break;
@@ -316,11 +358,11 @@ class CallFinder
                     }
                 }
 
-                $param = array(
+                $param = [
                     'name' => self::tokensToString($name),
                     'path' => self::tokensToString(self::tokensTrim($param['full'])),
                     'expression' => $expression,
-                );
+                ];
             }
 
             // Get the modifiers
@@ -334,7 +376,7 @@ class CallFinder
                 --$index;
             }
 
-            $mods = array();
+            $mods = [];
 
             while (isset($tokens[$index])) {
                 if (isset(self::$ignore[$tokens[$index][0]])) {
@@ -351,10 +393,10 @@ class CallFinder
                 break;
             }
 
-            $function_calls[] = array(
+            $function_calls[] = [
                 'parameters' => $params,
                 'modifiers' => $mods,
-            );
+            ];
         }
 
         return $function_calls;
@@ -430,10 +472,11 @@ class CallFinder
     private static function tokensFormatted(array $tokens)
     {
         $space = false;
+        $attribute = false;
 
         $tokens = self::tokensTrim($tokens);
 
-        $output = array();
+        $output = [];
         $last = null;
 
         foreach ($tokens as $index => $token) {
@@ -444,7 +487,10 @@ class CallFinder
 
                 $next = $tokens[self::realTokenIndex($tokens, $index)];
 
-                if (isset(self::$strip[$last[0]]) && !self::tokenIsOperator($next)) {
+                /** @var array|string $last */
+                if ($attribute && ']' === $last[0]) {
+                    $attribute = false;
+                } elseif (isset(self::$strip[$last[0]]) && !self::tokenIsOperator($next)) {
                     continue;
                 }
 
@@ -455,6 +501,10 @@ class CallFinder
                 $token = ' ';
                 $space = true;
             } else {
+                if (KINT_PHP80 && $last && T_ATTRIBUTE == $last[0]) {
+                    $attribute = true;
+                }
+
                 $space = false;
                 $last = $token;
             }
