@@ -25,8 +25,8 @@
 
 namespace Kint\Parser;
 
-use Kint\Object\BasicObject;
-use Kint\Object\InstanceObject;
+use Kint\Zval\InstanceValue;
+use Kint\Zval\Value;
 
 class BlacklistPlugin extends Plugin
 {
@@ -35,32 +35,18 @@ class BlacklistPlugin extends Plugin
      *
      * @var array
      */
-    public static $blacklist = array();
+    public static $blacklist = [];
 
     /**
      * List of classes and interfaces to blacklist except when dumped directly.
      *
      * @var array
      */
-    public static $shallow_blacklist = array();
-
-    /**
-     * Maximum size of arrays before blacklisting.
-     *
-     * @var int
-     */
-    public static $array_limit = 10000;
-
-    /**
-     * Maximum size of arrays before blacklisting except when dumped directly.
-     *
-     * @var int
-     */
-    public static $shallow_array_limit = 1000;
+    public static $shallow_blacklist = ['Psr\\Container\\ContainerInterface'];
 
     public function getTypes()
     {
-        return array('object', 'array');
+        return ['object'];
     }
 
     public function getTriggers()
@@ -68,21 +54,11 @@ class BlacklistPlugin extends Plugin
         return Parser::TRIGGER_BEGIN;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
-    {
-        if (\is_object($var)) {
-            return $this->parseObject($var, $o);
-        }
-        if (\is_array($var)) {
-            return $this->parseArray($var, $o);
-        }
-    }
-
-    protected function parseObject(&$var, BasicObject &$o)
+    public function parse(&$var, Value &$o, $trigger)
     {
         foreach (self::$blacklist as $class) {
             if ($var instanceof $class) {
-                return $this->blacklistObject($var, $o);
+                return $this->blacklistValue($var, $o);
             }
         }
 
@@ -92,48 +68,20 @@ class BlacklistPlugin extends Plugin
 
         foreach (self::$shallow_blacklist as $class) {
             if ($var instanceof $class) {
-                return $this->blacklistObject($var, $o);
+                return $this->blacklistValue($var, $o);
             }
         }
     }
 
-    protected function blacklistObject(&$var, BasicObject &$o)
+    protected function blacklistValue(&$var, Value &$o)
     {
-        $object = new InstanceObject();
+        $object = new InstanceValue();
         $object->transplant($o);
         $object->classname = \get_class($var);
-        $object->hash = \spl_object_hash($var);
+        $object->spl_object_hash = \spl_object_hash($var);
         $object->clearRepresentations();
         $object->value = null;
         $object->size = null;
-        $object->hints[] = 'blacklist';
-
-        $o = $object;
-
-        $this->parser->haltParse();
-    }
-
-    protected function parseArray(array &$var, BasicObject &$o)
-    {
-        if (\count($var) > self::$array_limit) {
-            return $this->blacklistArray($var, $o);
-        }
-
-        if ($o->depth <= 0) {
-            return;
-        }
-
-        if (\count($var) > self::$shallow_array_limit) {
-            return $this->blacklistArray($var, $o);
-        }
-    }
-
-    protected function blacklistArray(array &$var, BasicObject &$o)
-    {
-        $object = new BasicObject();
-        $object->transplant($o);
-        $object->value = null;
-        $object->size = \count($var);
         $object->hints[] = 'blacklist';
 
         $o = $object;
