@@ -25,19 +25,19 @@
 
 namespace Kint\Parser;
 
-use Kint\Object\BasicObject;
-use Kint\Object\InstanceObject;
-use Kint\Object\Representation\Representation;
+use Kint\Zval\InstanceValue;
+use Kint\Zval\Representation\Representation;
+use Kint\Zval\Value;
 use ReflectionClass;
 use ReflectionProperty;
 
 class ClassStaticsPlugin extends Plugin
 {
-    private static $cache = array();
+    private static $cache = [];
 
     public function getTypes()
     {
-        return array('object');
+        return ['object'];
     }
 
     public function getTriggers()
@@ -45,7 +45,7 @@ class ClassStaticsPlugin extends Plugin
         return Parser::TRIGGER_SUCCESS;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
+    public function parse(&$var, Value &$o, $trigger)
     {
         $class = \get_class($var);
         $reflection = new ReflectionClass($class);
@@ -53,14 +53,14 @@ class ClassStaticsPlugin extends Plugin
         // Constants
         // TODO: PHP 7.1 allows private consts but reflection doesn't have a way to check them yet
         if (!isset(self::$cache[$class])) {
-            $consts = array();
+            $consts = [];
 
             foreach ($reflection->getConstants() as $name => $val) {
-                $const = BasicObject::blank($name, '\\'.$class.'::'.$name);
+                $const = Value::blank($name, '\\'.$class.'::'.$name);
                 $const->const = true;
                 $const->depth = $o->depth + 1;
                 $const->owner_class = $class;
-                $const->operator = BasicObject::OPERATOR_STATIC;
+                $const->operator = Value::OPERATOR_STATIC;
                 $const = $this->parser->parse($val, $const);
 
                 $consts[] = $const;
@@ -73,18 +73,18 @@ class ClassStaticsPlugin extends Plugin
         $statics->contents = self::$cache[$class];
 
         foreach ($reflection->getProperties(ReflectionProperty::IS_STATIC) as $static) {
-            $prop = new BasicObject();
+            $prop = new Value();
             $prop->name = '$'.$static->getName();
             $prop->depth = $o->depth + 1;
             $prop->static = true;
-            $prop->operator = BasicObject::OPERATOR_STATIC;
+            $prop->operator = Value::OPERATOR_STATIC;
             $prop->owner_class = $static->getDeclaringClass()->name;
 
-            $prop->access = BasicObject::ACCESS_PUBLIC;
+            $prop->access = Value::ACCESS_PUBLIC;
             if ($static->isProtected()) {
-                $prop->access = BasicObject::ACCESS_PROTECTED;
+                $prop->access = Value::ACCESS_PROTECTED;
             } elseif ($static->isPrivate()) {
-                $prop->access = BasicObject::ACCESS_PRIVATE;
+                $prop->access = Value::ACCESS_PRIVATE;
             }
 
             if ($this->parser->childHasPath($o, $prop)) {
@@ -92,31 +92,37 @@ class ClassStaticsPlugin extends Plugin
             }
 
             $static->setAccessible(true);
-            $static = $static->getValue();
-            $statics->contents[] = $this->parser->parse($static, $prop);
+
+            if (KINT_PHP74 && !$static->isInitialized()) {
+                $prop->type = 'uninitialized';
+                $statics->contents[] = $prop;
+            } else {
+                $static = $static->getValue();
+                $statics->contents[] = $this->parser->parse($static, $prop);
+            }
         }
 
         if (empty($statics->contents)) {
             return;
         }
 
-        \usort($statics->contents, array('Kint\\Parser\\ClassStaticsPlugin', 'sort'));
+        \usort($statics->contents, ['Kint\\Parser\\ClassStaticsPlugin', 'sort']);
 
         $o->addRepresentation($statics);
     }
 
-    private static function sort(BasicObject $a, BasicObject $b)
+    private static function sort(Value $a, Value $b)
     {
         $sort = ((int) $a->const) - ((int) $b->const);
         if ($sort) {
             return $sort;
         }
 
-        $sort = BasicObject::sortByAccess($a, $b);
+        $sort = Value::sortByAccess($a, $b);
         if ($sort) {
             return $sort;
         }
 
-        return InstanceObject::sortByHierarchy($a->owner_class, $b->owner_class);
+        return InstanceValue::sortByHierarchy($a->owner_class, $b->owner_class);
     }
 }
