@@ -25,17 +25,17 @@
 
 namespace Kint\Parser;
 
-use Kint\Object\BasicObject;
-use Kint\Object\Representation\SplFileInfoRepresentation;
+use Kint\Zval\Representation\SplFileInfoRepresentation;
+use Kint\Zval\Value;
 use SplFileInfo;
 
 class FsPathPlugin extends Plugin
 {
-    public static $blacklist = array('/', '.');
+    public static $blacklist = ['/', '.'];
 
     public function getTypes()
     {
-        return array('string');
+        return ['string'];
     }
 
     public function getTriggers()
@@ -43,17 +43,17 @@ class FsPathPlugin extends Plugin
         return Parser::TRIGGER_SUCCESS;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
+    public function parse(&$var, Value &$o, $trigger)
     {
         if (\strlen($var) > 2048) {
             return;
         }
 
-        if (\preg_match('/[:?<>"*|]/', $var)) {
+        if (!\preg_match('/[\\/\\'.DIRECTORY_SEPARATOR.']/', $var)) {
             return;
         }
 
-        if (!\preg_match('/[\\/\\.\\'.DIRECTORY_SEPARATOR.']/', $var)) {
+        if (\preg_match('/[?<>"*|]/', $var)) {
             return;
         }
 
