@@ -25,18 +25,19 @@
 
 namespace Kint\Parser;
 
-use Kint\Object\BasicObject;
-use Kint\Object\TraceFrameObject;
-use Kint\Object\TraceObject;
 use Kint\Utils;
+use Kint\Zval\TraceFrameValue;
+use Kint\Zval\TraceValue;
+use Kint\Zval\Value;
 
 class TracePlugin extends Plugin
 {
-    public static $blacklist = array('spl_autoload_call');
+    public static $blacklist = ['spl_autoload_call'];
+    public static $path_blacklist = [];
 
     public function getTypes()
     {
-        return array('array');
+        return ['array'];
     }
 
     public function getTriggers()
@@ -44,27 +45,29 @@ class TracePlugin extends Plugin
         return Parser::TRIGGER_SUCCESS;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
+    public function parse(&$var, Value &$o, $trigger)
     {
         if (!$o->value) {
             return;
         }
 
+        /** @var array[] $trace Psalm workaround */
         $trace = $this->parser->getCleanArray($var);
 
         if (\count($trace) !== \count($o->value->contents) || !Utils::isTrace($trace)) {
             return;
         }
 
-        $traceobj = new TraceObject();
+        $traceobj = new TraceValue();
         $traceobj->transplant($o);
         $rep = $traceobj->value;
 
         $old_trace = $rep->contents;
 
         Utils::normalizeAliases(self::$blacklist);
+        $path_blacklist = self::normalizePaths(self::$path_blacklist);
 
-        $rep->contents = array();
+        $rep->contents = [];
 
         foreach ($old_trace as $frame) {
             $index = $frame->name;
@@ -78,7 +81,16 @@ class TracePlugin extends Plugin
                 continue;
             }
 
-            $rep->contents[$index] = new TraceFrameObject($frame, $trace[$index]);
+            if (isset($trace[$index]['file'])) {
+                $realfile = \realpath($trace[$index]['file']);
+                foreach ($path_blacklist as $path) {
+                    if (0 === \strpos($realfile, $path)) {
+                        continue 2;
+                    }
+                }
+            }
+
+            $rep->contents[$index] = new TraceFrameValue($frame, $trace[$index]);
         }
 
         \ksort($rep->contents);
@@ -88,5 +100,21 @@ class TracePlugin extends Plugin
         $traceobj->addRepresentation($rep);
         $traceobj->size = \count($rep->contents);
         $o = $traceobj;
+    }
+
+    protected static function normalizePaths(array $paths)
+    {
+        $normalized = [];
+
+        foreach ($paths as $path) {
+            $realpath = \realpath($path);
+            if (\is_dir($realpath)) {
+                $realpath .= DIRECTORY_SEPARATOR;
+            }
+
+            $normalized[] = $realpath;
+        }
+
+        return $normalized;
     }
 }
