@@ -26,46 +26,14 @@
 namespace Kint\Test;
 
 use Kint\Utils;
-use PHPUnit\Framework\TestCase;
 use ReflectionMethod;
+use ReflectionParameter;
 
-class UtilsTest extends TestCase
+class UtilsTest extends KintTestCase
 {
     protected $composer_stash;
     protected $installed_stash;
     protected $composer_test_dir;
-
-    protected function setUp()
-    {
-        parent::setUp();
-
-        if (\getenv('KINT_FILE')) {
-            $this->composer_test_dir = \dirname(__DIR__);
-        } else {
-            $this->composer_test_dir = KINT_DIR;
-        }
-
-        $this->composer_stash = \file_get_contents($this->composer_test_dir.'/composer.json');
-        $this->installed_stash = \file_get_contents($this->composer_test_dir.'/vendor/composer/installed.json');
-    }
-
-    protected function tearDown()
-    {
-        parent::tearDown();
-
-        if ($this->composer_stash) {
-            \file_put_contents($this->composer_test_dir.'/composer.json', $this->composer_stash);
-            \file_put_contents($this->composer_test_dir.'/vendor/composer/installed.json', $this->installed_stash);
-            $this->composer_stash = null;
-            $this->installed_stash = null;
-            if (\file_exists($this->composer_test_dir.'/composer/installed.json')) {
-                \unlink($this->composer_test_dir.'/composer/installed.json');
-            }
-            if (\file_exists($this->composer_test_dir.'/composer')) {
-                \rmdir($this->composer_test_dir.'/composer');
-            }
-        }
-    }
 
     public function testConstruct()
     {
@@ -75,64 +43,99 @@ class UtilsTest extends TestCase
 
     public function humanReadableBytesProvider()
     {
-        return array(
-            '1 byte' => array(
+        return [
+            '-1 byte' => [
+                -1,
+                [
+                    'value' => -1.0,
+                    'unit' => 'B',
+                ],
+            ],
+            '0 bytes' => [
+                0,
+                [
+                    'value' => 0.0,
+                    'unit' => 'B',
+                ],
+            ],
+            '1 byte' => [
                 1,
-                array(
+                [
                     'value' => 1.0,
                     'unit' => 'B',
-                ),
-            ),
-            '1023 bytes' => array(
+                ],
+            ],
+            '<1 k' => [
                 1023,
-                array(
+                [
                     'value' => 1023.0,
                     'unit' => 'B',
-                ),
-            ),
-            '1024 bytes' => array(
+                ],
+            ],
+            '1 k' => [
                 1024,
-                array(
+                [
                     'value' => 1.0,
                     'unit' => 'KB',
-                ),
-            ),
-            '1 meg' => array(
-                \pow(2, 20),
-                array(
+                ],
+            ],
+            '<1 meg' => [
+                \pow(2, 20) - 1,
+                [
                     'value' => 1.0,
                     'unit' => 'MB',
-                ),
-            ),
-            '1 gig' => array(
-                \pow(2, 30),
-                array(
+                ],
+            ],
+            '1 meg' => [
+                \pow(2, 20),
+                [
+                    'value' => 1.0,
+                    'unit' => 'MB',
+                ],
+            ],
+            '<1 gig' => [
+                \pow(2, 30) - 1,
+                [
                     'value' => 1.0,
                     'unit' => 'GB',
-                ),
-            ),
-            '1 tig' => array(
-                \pow(2, 40),
-                array(
+                ],
+            ],
+            '1 gig' => [
+                \pow(2, 30),
+                [
+                    'value' => 1.0,
+                    'unit' => 'GB',
+                ],
+            ],
+            '<1 tig' => [
+                \pow(2, 40) - 1,
+                [
                     'value' => 1.0,
                     'unit' => 'TB',
-                ),
-            ),
-            '>1 tig' => array(
+                ],
+            ],
+            '1 tig' => [
+                \pow(2, 40),
+                [
+                    'value' => 1.0,
+                    'unit' => 'TB',
+                ],
+            ],
+            '>1 tig' => [
                 \pow(2, 50),
-                array(
+                [
                     'value' => 1024.0,
                     'unit' => 'TB',
-                ),
-            ),
-            'halfway' => array(
+                ],
+            ],
+            'halfway' => [
                 15000,
-                array(
-                    'value' => 15000 / 1024,
+                [
+                    'value' => \round(15000 / 1024, 1),
                     'unit' => 'KB',
-                ),
-            ),
-        );
+                ],
+            ],
+        ];
     }
 
     /**
@@ -147,50 +150,74 @@ class UtilsTest extends TestCase
         $this->assertSame($expect, Utils::getHumanReadableBytes($input));
     }
 
-    public function sequentialArrayProvider()
+    public function arrayProvider()
     {
-        return array(
-            'sequential array' => array(
-                array(1, 2, 3),
+        return [
+            'sequential array' => [
+                [1, 2, 3],
                 true,
-            ),
-            'explicit sequential array' => array(
-                array(0 => 1, 1 => 2, 2 => 3),
+                false,
+            ],
+            'explicit sequential array' => [
+                [0 => 1, 1 => 2, 2 => 3],
                 true,
-            ),
-            'arrays start at 1' => array(
-                array(1 => 1, 2 => 2, 3 => 3),
                 false,
-            ),
-            'wrong order' => array(
-                array(0 => 1, 2 => 2, 1 => 3),
+            ],
+            'arrays start at 1' => [
+                [1 => 1, 2 => 2, 3 => 3],
                 false,
-            ),
-            'string keys' => array(
-                array(0 => 1, 1 => 2, 'two' => 3),
                 false,
-            ),
-            'string int keys' => array(
-                array('0' => 1, '1' => 2, '2' => 3),
+            ],
+            'wrong order' => [
+                [0 => 1, 2 => 2, 1 => 3],
+                false,
+                false,
+            ],
+            'string keys' => [
+                [0 => 1, 1 => 2, 'two' => 3],
+                false,
                 true,
-            ),
-            'padded string int keys' => array(
-                array('00' => 1, '01' => 2, '02' => 3),
+            ],
+            'string int keys' => [
+                ['0' => 1, '1' => 2, '2' => 3],
+                true,
                 false,
-            ),
-        );
+            ],
+            'padded string int keys' => [
+                ['00' => 1, '01' => 2, '02' => 3],
+                false,
+                true,
+            ],
+        ];
     }
 
     /**
      * @covers \Kint\Utils::isSequential
-     * @dataProvider sequentialArrayProvider
+     * @dataProvider arrayProvider
      *
      * @param array $input
-     * @param bool  $expect
+     * @param bool  $seq
      */
-    public function testIsSequential($input, $expect)
+    public function testIsSequential($input, $seq)
     {
-        $this->assertSame($expect, Utils::isSequential($input));
+        $this->assertSame($seq, Utils::isSequential($input));
+    }
+
+    /**
+     * @covers \Kint\Utils::isAssoc
+     * @dataProvider arrayProvider
+     *
+     * @param array $input
+     * @param bool  $_
+     * @param bool  $assoc
+     */
+    public function testIsAssoc($input, $_, $assoc)
+    {
+        $this->assertSame($assoc, Utils::isAssoc($input));
+
+        $input['key'] = 'val';
+
+        $this->assertTrue(Utils::isAssoc($input));
     }
 
     /**
@@ -201,37 +228,37 @@ class UtilsTest extends TestCase
      */
     public function testComposerGetExtras()
     {
-        \file_put_contents($this->composer_test_dir.'/composer.json', \json_encode(array(
-            'extra' => array(
-                'kint' => array('test' => 'data'),
-            ),
-        )));
+        \file_put_contents($this->composer_test_dir.'/composer.json', \json_encode([
+            'extra' => [
+                'kint' => ['test' => 'data'],
+            ],
+        ]));
 
-        if (\getenv('KINT_FILE')) {
-            $this->assertSame(array(), Utils::composerGetExtras('kint'));
+        if (\getenv('KINT_PHAR_TEST')) {
+            $this->assertSame([], Utils::composerGetExtras('kint'));
 
             return;
         }
 
-        $this->assertSame(array('test' => 'data'), Utils::composerGetExtras('kint'));
+        $this->assertSame(['test' => 'data'], Utils::composerGetExtras('kint'));
 
         \mkdir($this->composer_test_dir.'/composer');
         \unlink($this->composer_test_dir.'/vendor/composer/installed.json');
 
-        \file_put_contents($this->composer_test_dir.'/composer/installed.json', \json_encode(array(
-            array(
-                'extra' => array(
-                    'kint' => array('more' => 'test', 'data'),
-                ),
-            ),
-            array(
-                'extra' => array(
-                    'kint' => array('test' => 'ing'),
-                ),
-            ),
-        )));
+        \file_put_contents($this->composer_test_dir.'/composer/installed.json', \json_encode([
+            [
+                'extra' => [
+                    'kint' => ['more' => 'test', 'data'],
+                ],
+            ],
+            [
+                'extra' => [
+                    'kint' => ['test' => 'ing'],
+                ],
+            ],
+        ]));
 
-        $this->assertSame(array('more' => 'test', 'data', 'test' => 'ing'), Utils::composerGetExtras('kint'));
+        $this->assertSame(['more' => 'test', 'data', 'test' => 'ing'], Utils::composerGetExtras('kint'));
     }
 
     public function traceProvider()
@@ -242,46 +269,45 @@ class UtilsTest extends TestCase
         $bad_bt_2 = $bt;
         $bad_bt_2[0]['function'] = 1234;
 
-        return array(
-            'empty' => array(
-                'trace' => array(),
+        return [
+            'empty' => [
+                'trace' => [],
                 'expect' => false,
-            ),
-            'backtrace' => array(
+            ],
+            'backtrace' => [
                 'trace' => $bt,
                 'expect' => true,
-            ),
-            'bad backtrace, extra key' => array(
+            ],
+            'bad backtrace, extra key' => [
                 'trace' => $bad_bt_1,
                 'expect' => false,
-            ),
-            'bad backtrace, wrong type' => array(
+            ],
+            'bad backtrace, wrong type' => [
                 'trace' => $bad_bt_2,
                 'expect' => false,
-            ),
-            'mythical' => array(
-                'trace' => array(
-                    array(
+            ],
+            'mythical' => [
+                'trace' => [
+                    [
                         'function' => 'mythical_internal_function_with_no_args_that_results_in_a_backtrace',
                         'file' => __FILE__,
                         'line' => 1,
-                    ),
-                ),
+                    ],
+                ],
                 'expect' => true,
-            ),
-            'normal array' => array(
-                'trace' => array(1, 2, 3),
+            ],
+            'normal array' => [
+                'trace' => [1, 2, 3],
                 'expect' => false,
-            ),
-        );
+            ],
+        ];
     }
 
     /**
      * @dataProvider traceProvider
      * @covers \Kint\Utils::isTrace
      *
-     * @param array $trace
-     * @param bool  $expected
+     * @param bool $expected
      */
     public function testIsTrace(array $trace, $expected)
     {
@@ -290,54 +316,52 @@ class UtilsTest extends TestCase
 
     public function frameProvider()
     {
-        return array(
-            'function match' => array(
-                'frame' => array(
+        return [
+            'function match' => [
+                'frame' => [
                     'function' => 'testAWeirdFunctionName',
-                ),
-                'matches' => array('testaweirdfunctionname'),
+                ],
+                'matches' => ['testaweirdfunctionname'],
                 'expected' => true,
-            ),
-            'function no match denormalized' => array(
-                'frame' => array(
+            ],
+            'function no match denormalized' => [
+                'frame' => [
                     'function' => 'testAWeirdFunctionName',
-                ),
-                'matches' => array('testAWeirdFunctionName'),
+                ],
+                'matches' => ['testAWeirdFunctionName'],
                 'expected' => false,
-            ),
-            'function no match method' => array(
-                'frame' => array(
+            ],
+            'function no match method' => [
+                'frame' => [
                     'function' => 'testAWeirdFunctionName',
-                ),
-                'matches' => array(array('test', 'testaweirdfunctionname')),
+                ],
+                'matches' => [['test', 'testaweirdfunctionname']],
                 'expected' => false,
-            ),
-            'method no match function' => array(
-                'frame' => array(
+            ],
+            'method no match function' => [
+                'frame' => [
                     'function' => 'testAWeirdFunctionName',
                     'class' => 'test',
-                ),
-                'matches' => array('testAWeirdFunctionName'),
+                ],
+                'matches' => ['testAWeirdFunctionName'],
                 'expected' => false,
-            ),
-            'method match' => array(
-                'frame' => array(
+            ],
+            'method match' => [
+                'frame' => [
                     'function' => 'testAWeirdFunctionName',
                     'class' => 'test',
-                ),
-                'matches' => array(array('test', 'testaweirdfunctionname')),
+                ],
+                'matches' => [['test', 'testaweirdfunctionname']],
                 'expected' => true,
-            ),
-        );
+            ],
+        ];
     }
 
     /**
      * @dataProvider frameProvider
      * @covers \Kint\Utils::traceFrameIsListed
      *
-     * @param array $frame
-     * @param array $matches
-     * @param bool  $expected
+     * @param bool $expected
      */
     public function testTraceFrameIsListed(array $frame, array $matches, $expected)
     {
@@ -349,28 +373,122 @@ class UtilsTest extends TestCase
      */
     public function testNormalizeAliases()
     {
-        $input = array(
+        $input = [
             'CamelCaseFunction',
             'snake_case_function',
             'One_of_the_FunctionsMyColleaguesMADE__',
             'stringThatCan\'tBeAfunction',
             'another string that can not be a function',
-            array('clASs', 'meThod'),
-            array($this, 'meThod'),
-            array('a', 'b', 'c'),
-            array('\\big\\long\\class\\name', 'method'),
-        );
+            '\\NameSpaced\\Function',
+            ['clASs', 'meThod'],
+            [$this, 'meThod'],
+            ['a', 'b', 'c'],
+            ['\\big\\long\\class\\name', 'method'],
+        ];
 
-        $expected = array(
+        $expected = [
             'camelcasefunction',
             'snake_case_function',
             'one_of_the_functionsmycolleaguesmade__',
-            array('class', 'method'),
-            array('big\\long\\class\\name', 'method'),
-        );
+            'function',
+            ['class', 'method'],
+            ['big\\long\\class\\name', 'method'],
+        ];
 
         Utils::normalizeAliases($input);
 
         $this->assertSame($expected, $input);
+    }
+
+    public function truncateStringProvider()
+    {
+        return [
+            'standard string' => [
+                'input' => '123456789',
+                'expect' => '123...',
+                'length' => 6,
+            ],
+            'overclip' => [
+                'input' => '123456789',
+                'expect' => '12345...',
+                'length' => 8,
+            ],
+            'correct length' => [
+                'input' => '123456789',
+                'expect' => '123456789',
+                'length' => 9,
+            ],
+            'over length' => [
+                'input' => '123456789',
+                'expect' => '123456789',
+                'length' => 100,
+            ],
+            'under length' => [
+                'input' => '123456',
+                'expect' => '123',
+                'length' => 3,
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider truncateStringProvider
+     * @covers \Kint\Utils::truncateString
+     *
+     * @param string $input
+     * @param string $expect
+     * @param int    $length
+     */
+    public function testTruncateString($input, $expect, $length)
+    {
+        $this->assertSame($expect, Utils::truncateString($input, $length));
+    }
+
+    /**
+     * @covers \Kint\Utils::getTypeString
+     */
+    public function testGetTypeString()
+    {
+        if (!KINT_PHP70) {
+            $this->markTestSkipped('Not testing PHP7+ parameter type hints on PHP5');
+        }
+
+        $param = new ReflectionParameter(['Kint\\Test\\Fixtures\\TestClass', 'arrayHint'], 'x');
+        $this->assertSame('array', Utils::getTypeString($param->getType()));
+
+        if (KINT_PHP71) {
+            $param = new ReflectionParameter(['Kint\\Test\\Fixtures\\Php71TestClass', 'typeHints'], 'p1');
+            $this->assertSame('?int', Utils::getTypeString($param->getType()));
+
+            $param = new ReflectionParameter(['Kint\\Test\\Fixtures\\Php71TestClass', 'typeHints'], 'nullable');
+            $this->assertSame('?string', Utils::getTypeString($param->getType()));
+        }
+    }
+
+    protected function kintUp()
+    {
+        parent::kintUp();
+
+        $this->composer_test_dir = \dirname(__DIR__);
+        $this->composer_stash = \file_get_contents($this->composer_test_dir.'/composer.json');
+        $this->installed_stash = \file_get_contents($this->composer_test_dir.'/vendor/composer/installed.json');
+    }
+
+    protected function kintDown()
+    {
+        parent::kintDown();
+
+        if ($this->composer_stash) {
+            \file_put_contents($this->composer_test_dir.'/composer.json', $this->composer_stash);
+            \file_put_contents($this->composer_test_dir.'/vendor/composer/installed.json', $this->installed_stash);
+            $this->composer_stash = null;
+            $this->installed_stash = null;
+            if (\file_exists($this->composer_test_dir.'/composer/installed.json')) {
+                \unlink($this->composer_test_dir.'/composer/installed.json');
+            }
+            if (\file_exists($this->composer_test_dir.'/composer')) {
+                \rmdir($this->composer_test_dir.'/composer');
+            }
+        }
     }
 }
