@@ -1,11 +1,25 @@
 if (typeof window.kintPlain === 'undefined') {
-    window.kintPlain = (function() {
+    window.kintPlain = (function () {
         'use strict';
 
-        window.addEventListener('load', function() {
-            window.kintShared.dedupe(['style.kint-plain-style', 'script.kint-plain-script']);
-        });
+        var kintPlain = {
+            initLoad: function () {
+                kintPlain.style = window.kintShared.dedupe(
+                    'style.kint-plain-style',
+                    kintPlain.style
+                );
+                kintPlain.script = window.kintShared.dedupe(
+                    'script.kint-plain-script',
+                    kintPlain.script
+                );
+            },
 
-        return {};
+            style: null,
+            script: null,
+        };
+
+        return kintPlain;
     })();
 }
+
+window.kintShared.runOnce(window.kintPlain.initLoad);
