@@ -1,24 +1,32 @@
 if (typeof window.kintShared === 'undefined') {
-    window.kintShared = (function() {
+    window.kintShared = (function () {
         'use strict';
 
         var kintShared = {
-            dedupe: function(selectors) {
-                selectors.forEach(function(selector) {
-                    var found = false;
+            dedupe: function (selector, keep) {
+                [].forEach.call(document.querySelectorAll(selector), function (elem) {
+                    if (!keep || !keep.ownerDocument.contains(keep)) {
+                        keep = elem;
+                    }
 
-                    [].forEach.call(document.querySelectorAll(selector), function(elem) {
-                        if (found) {
-                            elem.parentNode.removeChild(elem);
-                        } else {
-                            found = true;
-                        }
-                    });
+                    if (elem !== keep) {
+                        elem.parentNode.removeChild(elem);
+                    }
                 });
+
+                return keep;
+            },
+
+            runOnce: function (cb) {
+                if (document.readyState === 'complete') {
+                    cb();
+                } else {
+                    window.addEventListener('load', cb);
+                }
             },
         };
 
-        window.addEventListener('click', function(e) {
+        window.addEventListener('click', function (e) {
             'use strict';
 
             // add ajax call to contact editor but prevent link default action
