@@ -26,13 +26,13 @@
 namespace Kint\Parser;
 
 use ArrayObject;
-use Kint\Object\BasicObject;
+use Kint\Zval\Value;
 
 class ArrayObjectPlugin extends Plugin
 {
     public function getTypes()
     {
-        return array('object');
+        return ['object'];
     }
 
     public function getTriggers()
@@ -40,7 +40,7 @@ class ArrayObjectPlugin extends Plugin
         return Parser::TRIGGER_BEGIN;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
+    public function parse(&$var, Value &$o, $trigger)
     {
         if (!$var instanceof ArrayObject) {
             return;
