@@ -25,14 +25,14 @@
 
 namespace Kint\Parser;
 
-use Kint\Object\BasicObject;
-use Kint\Object\Representation\Representation;
+use Kint\Zval\Representation\Representation;
+use Kint\Zval\Value;
 
 class JsonPlugin extends Plugin
 {
     public function getTypes()
     {
-        return array('string');
+        return ['string'];
     }
 
     public function getTriggers()
@@ -40,7 +40,7 @@ class JsonPlugin extends Plugin
         return Parser::TRIGGER_SUCCESS;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
+    public function parse(&$var, Value &$o, $trigger)
     {
         if (!isset($var[0]) || ('{' !== $var[0] && '[' !== $var[0])) {
             return;
@@ -54,7 +54,7 @@ class JsonPlugin extends Plugin
 
         $json = (array) $json;
 
-        $base_obj = new BasicObject();
+        $base_obj = new Value();
         $base_obj->depth = $o->depth;
 
         if ($o->access_path) {
