@@ -26,16 +26,16 @@
 namespace Kint\Parser;
 
 use Exception;
-use Kint\Object\BasicObject;
-use Kint\Object\Representation\SourceRepresentation;
-use Kint\Object\ThrowableObject;
+use Kint\Zval\Representation\SourceRepresentation;
+use Kint\Zval\ThrowableValue;
+use Kint\Zval\Value;
 use Throwable;
 
 class ThrowablePlugin extends Plugin
 {
     public function getTypes()
     {
-        return array('object');
+        return ['object'];
     }
 
     public function getTriggers()
@@ -43,15 +43,17 @@ class ThrowablePlugin extends Plugin
         return Parser::TRIGGER_SUCCESS;
     }
 
-    public function parse(&$var, BasicObject &$o, $trigger)
+    public function parse(&$var, Value &$o, $trigger)
     {
         if (!$var instanceof Exception && (!KINT_PHP70 || !$var instanceof Throwable)) {
             return;
         }
 
-        $throw = new ThrowableObject($var);
+        $throw = new ThrowableValue($var);
         $throw->transplant($o);
-        $throw->addRepresentation(new SourceRepresentation($var->getFile(), $var->getLine()), 0);
+        $r = new SourceRepresentation($var->getFile(), $var->getLine());
+        $r->showfilename = true;
+        $throw->addRepresentation($r, 0);
 
         $o = $throw;
     }
