@@ -25,7 +25,9 @@
 
 namespace Kint\Renderer;
 
-use Kint\Object\BasicObject;
+use Exception;
+use Kint\Zval\Value;
+use Throwable;
 
 class CliRenderer extends TextRenderer
 {
@@ -57,21 +59,52 @@ class CliRenderer extends TextRenderer
      */
     public static $min_terminal_width = 40;
 
+    /**
+     * Which stream to check for VT100 support on windows.
+     *
+     * null uses STDOUT if it's defined
+     *
+     * @var null|resource
+     */
+    public static $windows_stream = null;
+
     protected static $terminal_width = null;
 
     protected $windows_output = false;
+
+    protected $colors = false;
 
     public function __construct()
     {
         parent::__construct();
 
-        if (!self::$force_utf8) {
-            $this->windows_output = KINT_WIN;
+        if (!self::$force_utf8 && KINT_WIN) {
+            if (!KINT_PHP72) {
+                $this->windows_output = true;
+            } else {
+                $stream = self::$windows_stream;
+
+                if (!$stream && \defined('STDOUT')) {
+                    $stream = STDOUT;
+                }
+
+                if (!$stream) {
+                    $this->windows_output = true;
+                } else {
+                    $this->windows_output = !\sapi_windows_vt100_support($stream);
+                }
+            }
         }
 
         if (!self::$terminal_width) {
             if (!KINT_WIN && self::$detect_width) {
-                self::$terminal_width = \exec('tput cols');
+                try {
+                    self::$terminal_width = \exec('tput cols');
+                } catch (Exception $e) {
+                    self::$terminal_width = self::$default_width;
+                } catch (Throwable $t) {
+                    self::$terminal_width = self::$default_width;
+                }
             }
 
             if (self::$terminal_width < self::$min_terminal_width) {
@@ -79,12 +112,14 @@ class CliRenderer extends TextRenderer
             }
         }
 
+        $this->colors = $this->windows_output ? false : self::$cli_colors;
+
         $this->header_width = self::$terminal_width;
     }
 
     public function colorValue($string)
     {
-        if (!self::$cli_colors) {
+        if (!$this->colors) {
             return $string;
         }
 
@@ -93,7 +128,7 @@ class CliRenderer extends TextRenderer
 
     public function colorType($string)
     {
-        if (!self::$cli_colors) {
+        if (!$this->colors) {
             return $string;
         }
 
@@ -102,14 +137,14 @@ class CliRenderer extends TextRenderer
 
     public function colorTitle($string)
     {
-        if (!self::$cli_colors) {
+        if (!$this->colors) {
             return $string;
         }
 
         return "\x1b[36m".\str_replace("\n", "\x1b[0m\n\x1b[36m", $string)."\x1b[0m";
     }
 
-    public function renderTitle(BasicObject $o)
+    public function renderTitle(Value $o)
     {
         if ($this->windows_output) {
             return $this->utf8ToWindows(parent::renderTitle($o));
@@ -140,8 +175,8 @@ class CliRenderer extends TextRenderer
     protected function utf8ToWindows($string)
     {
         return \str_replace(
-            array('┌', '═', '┐', '│', '└', '─', '┘'),
-            array("\xda", "\xdc", "\xbf", "\xb3", "\xc0", "\xc4", "\xd9"),
+            ['┌', '═', '┐', '│', '└', '─', '┘'],
+            [' ', '=', ' ', '|', ' ', '-', ' '],
             $string
         );
     }
