@@ -25,8 +25,8 @@
 
 namespace Kint\Renderer\Rich;
 
-use Kint\Object\BasicObject;
 use Kint\Renderer\RichRenderer;
+use Kint\Zval\Value;
 
 abstract class Plugin implements PluginInterface
 {
@@ -40,10 +40,9 @@ abstract class Plugin implements PluginInterface
     /**
      * Renders a locked header.
      *
-     * @param BasicObject $o
-     * @param string      $content
+     * @param string $content
      */
-    public function renderLockedHeader(BasicObject $o, $content)
+    public function renderLockedHeader(Value $o, $content)
     {
         $header = '<dt class="kint-parent kint-locked">';
 
