@@ -25,11 +25,11 @@
 
 namespace Kint\Renderer\Rich;
 
-use Kint\Object\BasicObject;
+use Kint\Zval\Value;
 
-class RecursionPlugin extends Plugin implements ObjectPluginInterface
+class RecursionPlugin extends Plugin implements ValuePluginInterface
 {
-    public function renderObject(BasicObject $o)
+    public function renderValue(Value $o)
     {
         return '<dl>'.$this->renderLockedHeader($o, '<var>Recursion</var>').'</dl>';
     }
