@@ -25,22 +25,22 @@
 
 namespace Kint\Renderer\Rich;
 
-use Kint\Object\Representation\Representation;
-use Kint\Object\Representation\SourceRepresentation;
+use Kint\Zval\Representation\Representation;
+use Kint\Zval\Representation\SourceRepresentation;
 
 class SourcePlugin extends Plugin implements TabPluginInterface
 {
     public function renderTab(Representation $r)
     {
         if (!($r instanceof SourceRepresentation) || empty($r->source)) {
-            return false;
+            return;
         }
 
         $source = $r->source;
 
         // Trim empty lines from the start and end of the source
         foreach ($source as $linenum => $line) {
-            if (\trim($line) || $linenum === $r->line) {
+            if (\strlen(\trim($line)) || $linenum === $r->line) {
                 break;
             }
 
@@ -48,35 +48,32 @@ class SourcePlugin extends Plugin implements TabPluginInterface
         }
 
         foreach (\array_reverse($source, true) as $linenum => $line) {
-            if (\trim($line) || $linenum === $r->line) {
+            if (\strlen(\trim($line)) || $linenum === $r->line) {
                 break;
             }
 
             unset($source[$linenum]);
         }
 
-        $start = '';
-        $highlight = '';
-        $end = '';
+        $output = '';
 
         foreach ($source as $linenum => $line) {
-            if ($linenum < $r->line) {
-                $start .= $line."\n";
-            } elseif ($linenum === $r->line) {
-                $highlight = '<div class="kint-highlight">'.$this->renderer->escape($line).'</div>';
+            if ($linenum === $r->line) {
+                $output .= '<div class="kint-highlight">'.$this->renderer->escape($line)."\n".'</div>';
             } else {
-                $end .= $line."\n";
+                $output .= '<div>'.$this->renderer->escape($line)."\n".'</div>';
             }
         }
-
-        $output = $this->renderer->escape($start).$highlight.$this->renderer->escape(\substr($end, 0, -1));
 
         if ($output) {
             \reset($source);
 
-            $marker = '@@ '.((int) \key($source)).','.\count($source).' @@';
+            $data = '';
+            if ($r->showfilename) {
+                $data = ' data-kint-filename="'.$this->renderer->escape($r->filename).'"';
+            }
 
-            return '<pre class="kint-source" data-kint-sourcerange="'.$marker.'">'.$output.'</pre>';
+            return '<div><pre class="kint-source"'.$data.' style="counter-reset: kint-l '.((int) \key($source) - 1).';">'.$output.'</pre></div><div></div>';
         }
     }
 }
