@@ -25,9 +25,12 @@
 
 namespace Kint\Renderer\Rich;
 
-use Kint\Object\Representation\Representation;
+use Kint\Zval\Representation\Representation;
 
 interface TabPluginInterface extends PluginInterface
 {
-    public function renderTab(Representation $o);
+    /**
+     * @return null|string
+     */
+    public function renderTab(Representation $r);
 }
