@@ -25,12 +25,12 @@
 
 namespace Kint\Renderer\Text;
 
-use Kint\Object\BasicObject;
-use Kint\Object\Representation\MicrotimeRepresentation;
 use Kint\Renderer\PlainRenderer;
 use Kint\Renderer\Rich\MicrotimePlugin as RichPlugin;
 use Kint\Renderer\TextRenderer;
 use Kint\Utils;
+use Kint\Zval\Representation\MicrotimeRepresentation;
+use Kint\Zval\Value;
 
 class MicrotimePlugin extends Plugin
 {
@@ -45,12 +45,12 @@ class MicrotimePlugin extends Plugin
         }
     }
 
-    public function render(BasicObject $o)
+    public function render(Value $o)
     {
         $r = $o->getRepresentation('microtime');
 
         if (!$r instanceof MicrotimeRepresentation) {
-            return false;
+            return;
         }
 
         $out = '';
