@@ -25,8 +25,8 @@
 
 namespace Kint\Renderer\Text;
 
-use Kint\Object\BasicObject;
 use Kint\Renderer\TextRenderer;
+use Kint\Zval\Value;
 
 abstract class Plugin
 {
@@ -37,5 +37,8 @@ abstract class Plugin
         $this->renderer = $r;
     }
 
-    abstract public function render(BasicObject $o);
+    /**
+     * @return null|string
+     */
+    abstract public function render(Value $o);
 }
