@@ -25,77 +25,23 @@
 
 namespace Kint\Test;
 
-use Kint\Kint;
-use Kint\Object\BlobObject;
-use Kint\Object\Representation\ColorRepresentation;
-use Kint\Parser\BlacklistPlugin;
-use Kint\Renderer\TextRenderer;
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Util_InvalidArgumentHelper;
+use Prophecy\PhpUnit\ProphecyTrait;
 
-abstract class KintTestCase extends TestCase
+class KintTestCase extends TestCase
 {
-    protected $kint_statics;
-    protected $char_encodings;
-    protected $legacy_encodings;
-    protected $text_decorations;
-    protected $text_plugin_whitelist;
-    protected $color_map;
-    protected $blacklist_plugin_blacklist;
-    protected $blacklist_plugin_shallow_blacklist;
-    protected $blacklist_plugin_array_limit;
-    protected $blacklist_plugin_shallow_array_limit;
+    use KintTestTrait;
+    use ProphecyTrait;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
-
-        $this->kint_statics = Kint::getStatics();
-        $this->char_encodings = BlobObject::$char_encodings;
-        $this->legacy_encodings = BlobObject::$legacy_encodings;
-        $this->text_decorations = TextRenderer::$decorations;
-        $this->text_plugin_whitelist = TextRenderer::$parser_plugin_whitelist;
-        $this->color_map = ColorRepresentation::$color_map;
-        $this->blacklist_plugin_blacklist = BlacklistPlugin::$blacklist;
-        $this->blacklist_plugin_shallow_blacklist = BlacklistPlugin::$shallow_blacklist;
-        $this->blacklist_plugin_array_limit = BlacklistPlugin::$array_limit;
-        $this->blacklist_plugin_shallow_array_limit = BlacklistPlugin::$shallow_array_limit;
+        static::kintUp();
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
-
-        foreach ($this->kint_statics as $key => $val) {
-            Kint::${$key} = $val;
-        }
-
-        BlobObject::$char_encodings = $this->char_encodings;
-        BlobObject::$legacy_encodings = $this->legacy_encodings;
-        TextRenderer::$decorations = $this->text_decorations;
-        TextRenderer::$parser_plugin_whitelist = $this->text_plugin_whitelist;
-        ColorRepresentation::$color_map = $this->color_map;
-        BlacklistPlugin::$blacklist = $this->blacklist_plugin_blacklist;
-        BlacklistPlugin::$shallow_blacklist = $this->blacklist_plugin_shallow_blacklist;
-        BlacklistPlugin::$array_limit = $this->blacklist_plugin_array_limit;
-        BlacklistPlugin::$shallow_array_limit = $this->blacklist_plugin_shallow_array_limit;
-    }
-
-    /**
-     * Asserts that a condition is true.
-     *
-     * @param array  $expected
-     * @param string $actual
-     * @param string $message
-     *
-     * @throws PHPUnit_Framework_Exception
-     */
-    public function assertLike(array $expected, $actual, $message = '')
-    {
-        if (!\is_string($actual)) {
-            throw PHPUnit_Util_InvalidArgumentHelper::factory(2, 'string');
-        }
-
-        self::assertThat($actual, new ContainsInOrderConstraint($expected), $message);
+        static::kintDown();
     }
 }
